requirements.txt
numpy==1.26.4
pandas==2.2.2
keras==3.3.3
jupyter==1.0.0
joblib==1.4.2
tabulate~=0.9.0
telegram==0.0.1
matplotlib==3.9.0
tensorflow==2.16.1
paho-mqtt==2.1.0
scikit-learn==1.5.0
google-cloud==0.34.0
python-telegram-bot==21.2
google-cloud-storage==2.16.0
google-cloud-logging==3.10.0
scipy~=1.13.1
pathlib~=1.0.1
psycopg2~=2.9.9